# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Bolsa de Empleo",
        "description": "Sistema de gestión de ofertas de empleo y seguimiento de candidatos.",
        "contact": {
            "email": "admin@centro.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor de Desarrollo Local"
        }
    ],
    "paths": {
        "/api/admin/gestion": {
            "get": {
                "tags": [
                    "Gestión Admin"
                ],
                "summary": "Listar usuarios administradores",
                "operationId": "835bb8255873dffc9ddad49518e1abbb",
                "parameters": [
                    {
                        "name": "rows",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "example": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Listado obtenido"
                    },
                    "403": {
                        "description": "No es SuperAdmin"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Gestión Admin"
                ],
                "summary": "Crear nuevo administrador",
                "operationId": "5baabee64fcbe1988e5c26999e6f0ea6",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Admin creado con clave temporal"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/admin/gestion/{id}/reset-password": {
            "patch": {
                "tags": [
                    "Gestión Admin"
                ],
                "summary": "Resetear password de un administrador",
                "operationId": "32bef5ed85acad786a18c3c3cdf9f0a5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Password reseteado"
                    },
                    "403": {
                        "description": "Intento de resetear al SuperAdmin o acceso no permitido"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/api/registro": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Registro de nuevo usuario",
                "operationId": "33bd69a3842f1c227d32464488e95703",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password",
                                    "role"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Usuario Ejemplo"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "user@ejemplo.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "minLength": 6
                                    },
                                    "role": {
                                        "description": "2 para Empresa, 3 para Alumno",
                                        "type": "integer",
                                        "example": 3
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Usuario creado y token generado"
                    },
                    "409": {
                        "description": "Usuario inactivo (requiere reactivación)"
                    },
                    "422": {
                        "description": "Error de validación o email duplicado"
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Inicio de sesión",
                "operationId": "a3b306d14572d1f4bd6c064b3233e7b8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login exitoso con token y rol"
                    },
                    "401": {
                        "description": "Credenciales incorrectas"
                    },
                    "403": {
                        "description": "Cuenta inactiva o pendiente de validación"
                    }
                }
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Cierre de sesión",
                "operationId": "fe8f3429cd6979b3b4517e186505f9f9",
                "responses": {
                    "200": {
                        "description": "Token eliminado correctamente"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/perfil-auth": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Datos del usuario autenticado",
                "operationId": "8ba04db5dc69efab4a9effbe0aad00f8",
                "responses": {
                    "200": {
                        "description": "Datos de perfil"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Enter token in format (Bearer <token>)",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "name": "Authorization",
                "in": "header",
                "example": "Bearer 19|tBfCzyfhD1Ib3VA5HPyHfR6CjrIoyT5VPq7s0mMnb19cdfbe"
            }
        }
    },
...
}